requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
pdfplumber==0.11.9
chromadb==1.5.1
sentence-transformers==5.2.3
langchain==1.2.10
langchain-chroma==1.1.0
langchain-text-splitters==1.1.1
langchain-huggingface==1.2.0
google-api-python-client==2.190.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
gradio==6.6.0
openai==2.21.0